Added check for the existence of the path to the file for deletion.

+++ Gallery.Service/Services/ImageService.cs
@@ -141,7 +141,10 @@ namespace Gallery.Service
 
         public void Delete(string fileToDelete)
         {
-            File.Delete(fileToDelete);
+            if (File.Exists(fileToDelete))
+            {
+                File.Delete(fileToDelete);
+            }
         }
     }
 }
